Playing with Boost.Beast.
See handle_request function. There, I learned how to print the headers
of an HTTP request as well as other information.

boost::string_view's do not work with fmtlib. To fix this, I set a
Boost.Beast compilation option which instructs it to use
std::string_view instead.

This code is based on the following:

    https://www.boost.org/doc/libs/1_78_0/libs/beast/example/http/server/async/http_server_async.cpp