[#49] Added atomic operations for ACLs and metadata.
The changes in this commit removes the need for the /metadata endpoint.

This commit improves the API by moving the handling of metadata
modification operations into multiple endpoints. Those endpoints are:

    - /collections
    - /data-objects
    - /resources
    - /users-groups

This change also helps users understand/learn which entities within
iRODS support metadata.

This commit also adds support for modifying ACLs/permissions atomically.
The endpoints which support this are:

    - /collections
    - /data-objects