[#144] Add support for the anonymous iRODS user account.
When the anonymous user is detected, the HTTP API server will generate a
token allowing the client to proceed. No authentication is actually
performed. If a client attempts to execute an operation as the anonymous
user, but the zone doesn't contain the anonymous user account, the HTTP
API will return an error.

This commit also removes the password member variable from the
authenticated_client_info struct.