[irods/irods#7265,irods/irods#7434] Minor reorganization + CMake TLC
I've reorganized the project to follow the conventions of the main irods repo.
The individual endpoints seemed to be organized in such a way that they could
be expanded into multiple source files. I have preserved this in the reorg.
Additionally, a lot of the CMake was re-written. I stuck with object libraries
for everything in order to make it easier to set properties for groups of
source files.