[#223] Return HTTP 200 when inputs to stat operation are valid.
This change aligns the stat operation with how other operations are
designed. That is, the HTTP status code is for reporting issues specific
to the HTTP side of things.

Interactions with the iRODS server must be reported via the JSON
response.

At this time, only the read operation of the /data-objects endpoint is
allowed to communicate iRODS issues via the HTTP status code.