[#224] Skip Basic authentication test if OIDC is enabled.
When OIDC is enabled, the authentication test fails. This is because the
code path in the HTTP API server changes, resulting in a different HTTP
status code.

This commit allows clients to ask whether the HTTP API server has OIDC
enabled and react accordingly. This addition allows the test to skip
itself when necessary.