[#196] Distribute long-running write operations across thread pool.
Before this commit, it was possible for the server to become
unresponsive if enough clients executed long running write operations.

This was possible because of the following:
- The HTTP API uses a fixed number of threads to process client requests
- The handler for the write operation contains a for-loop which makes
  multiple calls to the iRODS server

This commit resolves the issue by sequentially posting writes to the
background thread pool. This change protects the HTTP API from spending
too much time on one client request.