[#343] Validate JWT Access Tokens locally
This PR implements the OAuth 2.0 Access Token JWT Profile to allow
for validation of JWT Access Tokens on the HTTP API.

Of minor note, this does not strictly follow the RFC as is. This is
because some OpenID Providers do not provide a 'typ' header value
as is specified in the RFC. This should not affect processing
otherwise.