[#338] OIDC: Confirm mapped iRODS username exists before returning token.
Before this change, the HTTP API would incorrectly return a token to the
end-user even though the mapped iRODS username didn't exist in the iRODS
catalog.

With this commit, the HTTP API checks the iRODS catalog for the mapped
iRODS username before returning a token. If the user does not exist, a
HTTP status code of "401 Unauthorized" is returned.