[#107] Validate OpenID ID Tokens in client mode
This work is largely based off of the local JWT validation work
in commit 4d0a5d6. Some changes in the validation function were
made to accommodate both types of tokens.

This commit also allows for non-compliance with the
OpenID Connect 1.0 specification, specifically for the secret key
for the symmetric ID Token keys.