[#375] Make strict introspection endpoint `aud` check opt-in
This commit makes the strict `aud` check for the introspection
endpoint optional. This is due to the fact that some OpenID
Providers may be configured to not provide an `aud` in the
endpoint response, even if the given access token contains the
information.

Additionally, documentation of this feature is added, along with
relevant schema checks.